(core) replace material in bower_components with package.json lines + yarn cache
Summary:
In preparation for moving the client code to grist-core, this strips out the third party code checked in to bower_components. Since bower maintainers are recommending migration, I list the needed packages in package.json.

 * bootstrap-tokenfield is removed
 * jqueryui is a later version (older version is not available on npmjs)

There is a theme image that was included by `electron-files.rsync` that is no longer present and does not seem to be needed?  I've trimmed material that no longer seems present from this file.

I've retained a `bower_components` directory with soft links to the needed material. The directory could be eliminated with some more work, but that would be easier to review as a separate diff. Packaging appears to work correctly, but it will be important to check on staging also.

Test Plan: existing tests pass; packaging spot checks look sane - client loads ok and can't tickle any 404s

Reviewers: dsagal

Reviewed By: dsagal

Differential Revision: https://phab.getgrist.com/D2623